pyproject.toml
[project]
name = "ai-pipeline-template"
version = "1.0.0"
description = "AI pipeline template using Claude for campaign name standardization"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.40.0",
    "google-cloud-bigquery>=3.25.0",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[tool.hatch.envs.default]
type = "virtual"
path = ".venv"
installer = "uv"

[tool.hatch.envs.default.scripts]
load = "python src/load_sample.py"
pipeline = "python src/pipeline.py"
analyze = "python src/analysis.py"
